Move components demo stuff into its own wizard
