+++ includes/class-newspack.php
@@ -60,7 +60,7 @@ final class Newspack {
 		include_once NEWSPACK_ABSPATH . 'includes/class-admin-plugins-screen.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-api.php';
 
-		include_once NEWSPACK_ABSPATH . '/includes/wizards/class-subscriptions-wizard.php';
+		include_once NEWSPACK_ABSPATH . '/includes/wizards/class-components-demo.php';
 	}
 
 	/**
